test(use-value): recategorize and clean up browser tests
